+++ pytest_django/plugin.py
@@ -399,12 +399,36 @@ def _django_setup_unittest(request, django_db_blocker):
         request.addfinalizer(teardown)
 
 
-@pytest.fixture(autouse=True, scope='function')
-def _django_clear_outbox(django_test_environment):
-    """Clear the django outbox, internal to pytest-django."""
-    if django_settings_is_configured():
-        from django.core import mail
-        del mail.outbox[:]
+class _DirectMailboxAccessProtector(list):
+
+    def _raise_assertion(*args, **kwargs):
+        raise AssertionError('To access mail.outbox, use the mailoutbox fixture.')
+
+    __len__ = __getitem__ = __nonzero__ = __bool__ = _raise_assertion
+
+
+@pytest.fixture(autouse=True)
+def _error_on_direct_mail_outbox_access(monkeypatch):
+    if not django_settings_is_configured():
+        return
+
+    from django.core import mail
+
+    outbox = _DirectMailboxAccessProtector()
+    monkeypatch.setattr(mail, 'outbox', outbox)
+    return outbox
+
+
+@pytest.fixture(scope='function')
+def mailoutbox(monkeypatch, _error_on_direct_mail_outbox_access):
+    if not django_settings_is_configured():
+        return
+
+    from django.core import mail
+
+    outbox = list()
+    monkeypatch.setattr(mail, 'outbox', outbox)
+    return outbox
 
 
 @pytest.fixture(autouse=True, scope='function')